ERD
Explanation:
Reservation table for storing contact information. (Can be used for tickets via email)
Movie table contains list of movies and their information (title, director, time,..)
Movie table can be updated to add/remove/update movies.
Screening table for times in different auditoriums.
With the Reservation table, we can then book a ticket for screenings.
The reservation table can store the email for ticket delivery in reservation_contact.
The reservation and seat reserverd tables keep track of seat that have been sold or reserved for each of the screenings.
The movie table can be updated regulary to reflect current movies avaible for screening.
And the auditoriu, table allows addittionaly new screens or new auditoriums for as the cinema expands.